src/main/java/com/anacleto/budgetcontrol/api/service/PessoaService.java
package com.anacleto.budgetcontrol.api.service;

import java.util.Optional;

import org.springframework.beans.BeanUtils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.stereotype.Service;

import com.anacleto.budgetcontrol.api.model.Pessoa;
import com.anacleto.budgetcontrol.api.repository.PessoaRepository;

@Service
public class PessoaService {
	
	@Autowired
	private PessoaRepository pessoaRepository;

	public Pessoa atualizar(Long codigo, Pessoa pessoa) {
		Optional<Pessoa> pessoaOp = pessoaRepository.findById(codigo);
		
		if (!pessoaOp.isPresent()) {
			throw new EmptyResultDataAccessException(1);
		}
		
		Pessoa pessoaSalva = pessoaOp.get();
		
		BeanUtils.copyProperties(pessoa, pessoaSalva, "codigo");
		
		return pessoaRepository.save(pessoaSalva);
	}
}
